+++ packages/remax-one/src/hostComponents/Button/index.ts
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import createHostComponent from '../../createHostComponent';
 import { TapEvent } from '../../types';
-import * as props from './props';
+import alias from './props/alias';
+import defaults from './props/default';
 
 export interface ButtonProps extends React.AriaAttributes {
   // 通用属性
@@ -24,6 +25,6 @@ export interface ButtonProps extends React.AriaAttributes {
   onTap?: (event: TapEvent) => void;
 }
 
-const Button = createHostComponent<ButtonProps>('button', props.alias, props.defaults);
+const Button = createHostComponent<ButtonProps>('button', alias, defaults);
 
 export default Button;
