+++ packages/remax-one/src/hostComponents/Input/props/index.ts
@@ -1,8 +1,6 @@
 import * as React from 'react';
 import { InputEvent } from '../../../types';
 
-export { default as alias } from './alias';
-
 interface CommonProps {
   // 通用属性
   readonly dataset?: DOMStringMap;
